+++ packages/tolk/test/golden/cstyle/generate_actual.ml
@@ -314,7 +314,7 @@ type test_case = {
 
 let all_renderers =
   [
-    ("clang", Cstyle.clang_no_abi Gpu_target.X86_64);
+    ("clang", Cstyle.clang_no_abi);
     ("cuda", Cstyle.cuda Gpu_target.SM80);
     ("metal", Cstyle.metal (Gpu_target.Apple 7));
     ("opencl", Cstyle.opencl "");
@@ -406,7 +406,7 @@ let test_cases =
       name = "vectorize_index";
       prog = make_vectorize_index_scalarized ();
       backends =
-        [ ("clang", Cstyle.clang_no_abi Gpu_target.X86_64); ("cuda", Cstyle.cuda Gpu_target.SM80) ];
+        [ ("clang", Cstyle.clang_no_abi); ("cuda", Cstyle.cuda Gpu_target.SM80) ];
     };
   ]